pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = "Habit tracking project which helps in setting up, managing an tracking your daily habits and routines"
authors = [{name = "Wojciech Bednarz"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "click>=8.3.0",
    "faker>=38.2.0",
    "fastapi[standard]>=0.121.1",
    "httpx>=0.28.1",
    "pipdeptree>=2.30.0",
    "pre-commit>=4.3.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib>=0.3.0",
    "pydantic-settings[standard]>=2.12.0",
    "pydantic[email]>=2.12.4",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "redis>=7.1.0",
    "ruff>=0.14.4",
    "sqlalchemy[asyncio,mypy]>=2.0.44",
    "testcontainers[postgres,redis]>=4.14.0",
    "types-redis>=4.6.0.20241004",
    "uvicorn>=0.38.0",
    "urllib3>=2.6.3",
    "jinja2>=3.1.6",
    "weasyprint>=68.0",
    "aioboto3>=15.5.0",
    "moto[s3,ses,sqs]>=5.1.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.7.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
disable_error_code = "misc"
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = "src.core.habit_async"
disable_error_code = ["arg-type"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "redis.asyncio.*"
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = "redis.exceptions"
disable_error_code = ["import-untyped"]


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--verbose"
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.pylint.master]
extension-pkg-allow-list = ["sqlalchemy"]

[tool.pylint.messages_control]
disable = [
  "logging-fstring-interpolation",
  "import-error",
  "C0114",
  "not-callable"
]

[dependency-groups]
dev = [
    "anyio>=4.11.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.4",
]

[[tool.mypy.overrides]]
module = "tests.*"
check_untyped_defs = false
disallow_untyped_defs = false
strict = false
